fix: Copy all subfolders of TF modules
In some cases, it might be desirable to have subfolders
inside a Terraform recipe folder (e.g., for local modules).

The current implementation ignored these, breaking the
initialization of the copied Terraform modules due to
missing files.

This commit changes the logic in terraform_utils.py to
copy any subfolders (not starting with a leading dot,
like `.terraform`) of a Terraform module to the destination.